LayoutTests/fast/webgpu/regression/repro_294360-expected.txt

CONSOLE MESSAGE: 0
CONSOLE MESSAGE: no validation error
